requirements.txt
aiocron==1.8
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==22.2.0
charset-normalizer==3.1.0
croniter==1.3.14
discord.py==2.2.2
frozenlist==1.3.3
idna==3.4
multidict==6.0.4
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz-deprecation-shim==0.1.0.post0
six==1.16.0
tzdata==2023.3
tzlocal==4.3
yarl==1.8.2
